feat: harmonize create/update sticky note
